Moved the UI to it's own file
moved all of the Ui function to it's own file.
