+++ base_ui.py
@@ -1,0 +1,100 @@
+import bpy
+   
+# this contains the main layout for the Sketch to mesh program
+# to link up functions with the buttons
+# first create the operator 
+# find the panel you want the funciton in(this cannot be inside the main function since everything is not inside the main panel except for the other panels)
+# use ( row.operator("(the operator you want)", text="(the name of the button)"))
+# for now all of the button will create a cube
+
+class VIEW3D_PT_Sketch_To_Mesh_Panel(bpy.types.Panel):  
+    bl_space_type = "VIEW_3D"  
+    bl_region_type = "UI"  
+    bl_idname = "_PT_Sketch_To_Mesh_Main_Panel" 
+
+    bl_category = "S-T-M"  # Sidebar cName
+    bl_label = "Sketch-To-Mesh"  # found at the top of the Panel
+
+    def draw(self, context): 
+        layout = self.layout
+
+        
+class VIEW3D_PT_Sketch_To_Mesh_Views_Panel(bpy.types.Panel):  
+    bl_label = "View"
+    bl_idname = "_PT_Views"
+    bl_parent_id = "_PT_Sketch_To_Mesh_Main_Panel" 
+    bl_space_type = 'VIEW_3D'
+    bl_region_type = 'UI'
+
+    def draw(self, context):
+        layout = self.layout
+        row = layout.row()
+        row.prop(context.scene, "front_views_file_path", text="Front View")
+        row = layout.row()
+        row.prop(context.scene, "back_views_file_path", text="Back View")
+        row = layout.row()
+        row.prop(context.scene, "side_views_file_path", text="Side View")
+        row = layout.row()
+        row.operator("object.reset_selected_images", text="Reset Images")
+
+
+class VIEW3D_PT_Sketch_To_Mesh_Align_Views_Panel(bpy.types.Panel):  
+    bl_label = "Align_Location"
+    bl_idname = "_PT_AlignViews"
+    bl_parent_id = "_PT_Views"
+    bl_space_type = 'VIEW_3D'
+    bl_region_type = 'UI'
+
+    def draw(self, context):
+        layout = self.layout
+        row = layout.row()
+        row.operator("object.place_image_in_space", text="Align Image")
+
+
+class VIEW3D_PT_Sketch_To_Mesh_MeshSettings_Panel(bpy.types.Panel):  
+    bl_label = "MeshSettings"
+    bl_idname = "_PT_MeshSettings"
+    bl_parent_id = "_PT_Sketch_To_Mesh_Main_Panel"  # Set the parent panel ID
+    bl_space_type = 'VIEW_3D'
+    bl_region_type = 'UI'
+
+    def draw(self, context):
+        layout = self.layout
+        layout.label(text="Mesh Settings")
+        row = layout.row()
+        row.prop(context.scene, "poly_count_range", text="Poly Count", slider=True)
+        row = layout.row()
+        row.operator("mesh.primitive_cube_add", text="Regenerate Preview Mesh")
+        row = layout.row()
+        row.operator("mesh.primitive_cube_add", text="Export Preview")
+        row = layout.row()
+        row.prop(context.scene, "mesh_rating", text="Mesh Rating", slider=True)
+        row = layout.row()
+        layout.label(text="Save Mesh")
+        row = layout.row()
+        layout.prop(context.scene, "FileName_Input", text="")
+        row = layout.row()
+        row.operator("mesh.primitive_cube_add", text="Export Mesh")
+
+
+class VIEW3D_PT_Sketch_To_Mesh_Testing(bpy.types.Panel):  
+    bl_label = "Testing"
+    bl_idname = "_PT_Testing_Panel"
+    bl_parent_id = "_PT_Sketch_To_Mesh_Main_Panel"
+    bl_space_type = 'VIEW_3D'
+    bl_region_type = 'UI'
+
+    def draw(self, context):
+        layout = self.layout
+        row = layout.row()
+        row.operator("wm.test_connection_operator", text="Test Connection")
+        row = layout.row()
+        row.operator("wm.database_login_popup", text="Access Database")
+        row = layout.row()
+        row.operator("wm.prepare_image_operator", text="Test Image Prep")
+        row = layout.row()
+        row.operator("wm.save_file_to_db_operator", text="Save File to DB")
+        row = layout.row()
+        row.operator("wm.get_file_from_db_operator", text="Get File from DB")
+        row = layout.row()
+        row.operator("wm.delete_file_from_db_operator", text="Delete File from DB")
